Move test func & add comment
