+++ metagpt/actions/minecraft/control_primitives_context/__init__.py
@@ -1,18 +1,20 @@
-import pkg_resources
 import os
-import voyager.utils as U
+import metagpt.utils.minecraft as utils
+from metagpt.logs import logger
 
-
-def load_control_primitives_context(primitive_names=None):
-    package_path = pkg_resources.resource_filename("metagpt", "")
-    if primitive_names is None:
-        primitive_names = [
-            primitive[:-3]
-            for primitive in os.listdir(f"{package_path}/actions/minecraft/control_primitives_context")
-            if primitive.endswith(".js")
+def load_skills_code_context(skill_names=None):
+    skills_dir = os.path.dirname(os.path.abspath(__file__))
+    if skill_names is None:
+        skill_names = [
+            skill[:-3] for skill in os.listdir(f"{skills_dir}") if skill.endswith(".js")
         ]
-    primitives = [
-        U.load_text(f"{package_path}/actions/minecraft/control_primitives_context/{primitive_name}.js")
-        for primitive_name in primitive_names
+    skills = [
+        utils.load_text(os.path.join(skills_dir, f"{skill_name}.js"))
+        for skill_name in skill_names
     ]
-    return primitives
+    return skills
+
+
+if __name__ == "__main__":
+    logger.info(load_skills_code_context(["craftItem", "exploreUntil"]))
+
