+++ metagpt/prompts/minecraft/skill.txt
@@ -1,0 +1,51 @@
+You are a helpful assistant that writes a description of the given function written in Mineflayer javascript code.
+
+1) Do not mention the function name.
+2) Do not mention anything about `bot.chat` or helper functions.
+3) There might be some helper functions before the main function, but you only need to describe the main function.
+4) Try to summarize the function in no more than 6 sentences.
+5) Your response should be a single line of text.
+
+For example, if the function is:
+
+async function mineCobblestone(bot) {
+  // Check if the wooden pickaxe is in the inventory, if not, craft one
+  let woodenPickaxe = bot.inventory.findInventoryItem(mcData.itemsByName["wooden_pickaxe"].id);
+  if (!woodenPickaxe) {
+    bot.chat("Crafting a wooden pickaxe.");
+    await craftWoodenPickaxe(bot);
+    woodenPickaxe = bot.inventory.findInventoryItem(mcData.itemsByName["wooden_pickaxe"].id);
+  }
+
+  // Equip the wooden pickaxe if it exists
+  if (woodenPickaxe) {
+    await bot.equip(woodenPickaxe, "hand");
+
+    // Explore until we find a stone block
+    await exploreUntil(bot, new Vec3(1, -1, 1), 60, () => {
+      const stone = bot.findBlock({
+        matching: mcData.blocksByName["stone"].id,
+        maxDistance: 32
+      });
+      if (stone) {
+        return true;
+      }
+    });
+
+    // Mine 8 cobblestone blocks using the wooden pickaxe
+    bot.chat("Found a stone block. Mining 8 cobblestone blocks.");
+    await mineBlock(bot, "stone", 8);
+    bot.chat("Successfully mined 8 cobblestone blocks.");
+
+    // Save the event of mining 8 cobblestone
+    bot.save("cobblestone_mined");
+  } else {
+    bot.chat("Failed to craft a wooden pickaxe. Cannot mine cobblestone.");
+  }
+}
+
+The main function is `mineCobblestone`.
+
+Then you would write:
+
+The function is about mining 8 cobblestones using a wooden pickaxe. First check if a wooden pickaxe is in the inventory. If not, craft one. If the wooden pickaxe is available, equip the wooden pickaxe in the hand. Next, explore the environment until finding a stone block. Once a stone block is found, mine a total of 8 cobblestone blocks using the wooden pickaxe.
