+++ metagpt/utils/minecraft/__init__.py
@@ -5,5 +5,4 @@
 from .load_prompts import load_prompt
 from .json_utils import *
 from .file_utils import *
-from .load_skills_code_context import load_skills_code_context
-from .action_rsp_parser import parse_action_response
+from .action_rsp_parser import parse_js_code, parse_action_response
