--- metagpt/utils/minecraft/skills_code_context/.prettierrc.json
@@ -1,3 +1,0 @@
-{
-    "tabWidth": 4
-}
